+++ src/main/java/com/redhat/quarkus/mandrel/collector/report/model/ImageStatsCollection.java
@@ -25,6 +25,8 @@ import jakarta.inject.Inject;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.Tuple;
 import jakarta.transaction.Transactional;
+import jakarta.ws.rs.WebApplicationException;
+import jakarta.ws.rs.core.Response;
 
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
@@ -43,7 +45,26 @@ public class ImageStatsCollection {
     EntityManager em;
 
     @Transactional
-    public ImageStats add(ImageStats stat) {
+    public ImageStats add(ImageStats stat, String tag, Long runnerInfoId) {
+        if (tag != null) {
+            stat.setTag(tag);
+        }
+        if (runnerInfoId != null) {
+            // Override the RunnerInfo if the ID is provided.
+            final RunnerInfo runnerInfo = getSingleRunnerInfo(runnerInfoId);
+            if (runnerInfo == null) {
+                throw new WebApplicationException("RunnerInfo with ID " + runnerInfoId + " not found.",
+                        Response.Status.NOT_FOUND);
+            }
+            stat.setRunnerInfo(runnerInfo);
+        } else {
+            if (stat.getRunnerInfo() != null && stat.getRunnerInfo().id == null) {
+                // If the RunnerInfo is not persisted yet, do it now
+                // to avoid a "save the transient instance before flushing" exception.
+                stat.getRunnerInfo().setCreatedAt(new Date());
+                em.persist(stat.getRunnerInfo());
+            }
+        }
         stat.setCreatedAt(new Date());
         em.persist(stat);
         return stat;
@@ -78,13 +99,25 @@ public class ImageStatsCollection {
                 .collect(HashMap::new, (m, t) -> m.put(t.get(1, Long.class), t.get(0, String.class)), HashMap::putAll);
     }
 
+    public ImageStats[] lookup(String what, ImageStats.SearchableRunnerInfo where, boolean wildcard) {
+        final String q = "SELECT s FROM ImageStats s WHERE s.runnerInfo." + where.column + " " +
+                (wildcard ? "LIKE CONCAT('%',:" + where.column + ",'%')" : "= :" + where.column) + " ORDER BY s.imageName";
+        return em.createQuery(q, ImageStats.class)
+                .setParameter(where.column, what)
+                .getResultList().toArray(new ImageStats[0]);
+    }
+
     public ImageStats getSingle(long id) {
         return em.find(ImageStats.class, id);
     }
 
     @Transactional
     public ImageStats deleteOne(long id) {
-        ImageStats stat = getSingle(id);
+        final ImageStats stat = getSingle(id);
+        if (stat == null) {
+            throw new WebApplicationException("ImageStats with ID " + id + " not found.",
+                    Response.Status.NOT_FOUND);
+        }
         em.remove(stat);
         return stat;
     }
@@ -133,8 +166,37 @@ public class ImageStatsCollection {
             return null;
         }
         if (info != null) {
+            info.setCreatedAt(new Date());
+            em.persist(info);
             stat.setRunnerInfo(info);
         }
         return stat;
+    }
+
+    // RunnerInfo
+
+    public RunnerInfo getSingleRunnerInfo(long id) {
+        return em.find(RunnerInfo.class, id);
+    }
+
+    @Transactional
+    public RunnerInfo add(RunnerInfo runnerInfo) {
+        runnerInfo.setCreatedAt(new Date());
+        em.persist(runnerInfo);
+        return runnerInfo;
+    }
+
+    /**
+     * RunnerInfo is optional, so we don't delete the attached stats alongside with it.
+     */
+    @Transactional
+    public RunnerInfo deleteRunnerInfo(long id) {
+        final RunnerInfo r = getSingleRunnerInfo(id);
+        if (r == null) {
+            throw new WebApplicationException("RunnerInfo with ID " + id + " not found.",
+                    Response.Status.NOT_FOUND);
+        }
+        em.remove(r);
+        return r;
     }
 }
